minor changes and bug fixes

+++ sigma_el_hh_hh.py
@@ -16,7 +16,7 @@ from poly import poly
 
 def Sigma_EL_HH_HH( E, vis = 0):
     E = np.array([E])
-    _E = float(E)
+    _E = E.astype(float)
     # ensures 3.03e0 < _E < 1.01e4
     _E = np.maximum(_E, 3.03e0)     
     _E = np.minimum(_E, 1.01e4)
